implemented all custom routings in pylons.
the url.environ thing is kept the same. Now it assumes the pattern /{controller}/{action}/{id}. {action} and {id} are set to None if not present.
As a result, those custom routings in pylons, whose controllers and actions were assigned directly in config/routing.py, now go to a redirect first, so that url.environ will work properly according to the redirected url that has a /{controller}/{action}/{id} pattern.

For example, in Pylons, we had
    map.connect('/hamlet/index', controller='contents', action='removal_of_hamlet')
This rule assigns the controller and the action directly, and url.environ['pylons.routes_dict'] is set up automatically and globally by the pylons system in advance once the route matches, and that's the reason why base controller can setup url.environ correctly in advance before reaching the actual controller class and action method.

However, in Pyramid, I googled a lot but couldn't find any system-level mechanism to hook the controller and action name to the global url.environ,
which means if the templates want to use the correct url.environ['pylons.routes_dict'] info for custom urls like '/hamlet/index', the settings must be made in every controller action,
i.e. putting url.environ['pylons.routes_dict']['controllers'] = ... in the __init__ method of every controller class, and putting url.environ['pylons.routes_dict']['action']=... in every action method. It is painful.

To work around this problem, I finally decided to go back to the old naive way that assumes the controller name and action name can be parsed directly from the /{controller}/{action} url.
As to those custom url routings, just make a redirection to the "correct" url that has an orthodox pattern.